Keep copies of items in cache
Modifications to an object after handing it over to set or after
obtaining it from get no longer pollute the in-memory cache.

Item is kept in cache in serialized form which has the additional
benefit of losing non-serializable properties already in the local cache
instead of losing them after object is first retrieved back from S3.